Cast byte to int before converting to string

+++ src/byte.c
@@ -7,5 +7,5 @@
 #include <the/string.h>
 
 the_str_t the_byte_str (unsigned char self) {
-  return the_str_alloc(L"%u", self);
+  return the_str_alloc(L"%u", (unsigned int) self);
 }
